Added Exceptions pack
