--- src/Nox.Types.EntityFramework.vNext/Exceptions/NoxEntityFrameworkException.cs
@@ -1,19 +1,0 @@
-﻿namespace Nox.Types.EntityFramework.vNext.Exceptions
-{
-    public class NoxEntityFrameworkException : Exception
-    {
-        public NoxEntityFrameworkException()
-        {
-        }
-
-        public NoxEntityFrameworkException(string message)
-            : base(message)
-        {
-        }
-
-        public NoxEntityFrameworkException(string message, Exception inner)
-            : base(message, inner)
-        {
-        }
-    }
-}
