+++ src/Nox.Types.EntityFramework/Exceptions/NoxEntityFrameworkException.cs
@@ -1,0 +1,18 @@
+namespace Nox.Types.EntityFramework.Exceptions;
+
+public class NoxEntityFrameworkException : Exception
+{
+    public NoxEntityFrameworkException()
+    {
+    }
+
+    public NoxEntityFrameworkException(string message)
+        : base(message)
+    {
+    }
+
+    public NoxEntityFrameworkException(string message, Exception inner)
+        : base(message, inner)
+    {
+    }
+}
